split withdrawStaked function from the execution script
