+++ src/utils/withdraw-staked.ts
@@ -1,5 +1,4 @@
 import * as hre from "hardhat";
-import { IDBVersion } from "../deploy/db/mongo-adapter/types";
 import { getMongoAdapter } from "../deploy/db/mongo-adapter/get-adapter";
 import { znsNames } from "../deploy/missions/contracts/names";
 import { ZNSTreasuryPausable__factory } from "../../typechain/factories/contracts/zns-pausable/treasury";
@@ -50,31 +49,3 @@ export const withdrawStakedByGovernor = async ({
 
   return tx;
 };
-
-// call the above function with await properly below
-void (async () => {
-  try {
-    const token = process.env.WITHDRAW_TOKEN_ADDRESS;
-    const to = process.env.TREASURY_WITHDRAW_RECIPIENT;
-
-    if (!token || !to) {
-      throw new Error("TOKEN_ADDRESS environment variable is not set");
-    }
-
-    const tx = await withdrawStakedByGovernor({
-      token,
-      to,
-    });
-
-    console.log(`Withdrawal transaction successful: ${tx.hash}`);
-    process.exit(0);
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-  } catch (error : Error) {
-    console.error(
-      `Error withdrawing staked tokens: ${error.message}
-      ${error.stack}`
-    );
-    process.exit(1);
-  }
-})();
